Update BigNum.java
Went through the method of going through each individual character and adding them up for the plus function.

multiplication works but is very inefficient with big numbers, to the point that its useless. This method adds the first value, a number of times. Another multiplication method should be used for efficiency.